requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asn1crypto=1.2.0=py37_0
beautifulsoup4=4.8.1=py37_0
blas=1.0=mkl
blis=0.4.1=pypi_0
boto3=1.10.25=pypi_0
botocore=1.13.25=pypi_0
ca-certificates=2019.11.27=0
catalogue=0.0.8=pypi_0
certifi=2019.11.28=py37_0
cffi=1.13.2=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
click=7.0=pypi_0
coverage=4.5.4=pypi_0
cryptography=2.8=py37h7a1dbc1_0
cudatoolkit=10.1.243=h74a9793_0
cycler=0.10.0=pypi_0
cymem=2.0.3=pypi_0
dependency-injector=3.14.12=pypi_0
docutils=0.15.2=pypi_0
entrypoints=0.3=pypi_0
fastprogress=0.1.22=pypi_0
flake8=3.7.9=pypi_0
freetype=2.9.1=ha9979f8_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.8=pypi_0
intel-openmp=2019.4=245
jmespath=0.9.4=pypi_0
joblib=0.14.0=py_0
jpeg=9b=hb83a4c4_2
kiwisolver=1.1.0=py37ha925a31_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_0
matplotlib=3.1.2=pypi_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
mock=3.0.5=pypi_0
ninja=1.9.0=py37h74a9793_0
nose=1.3.7=pypi_0
numexpr=2.7.0=pypi_0
numpy=1.17.4=pypi_0
numpy-base=1.17.4=py37hc3f5095_0
nvidia-ml-py3=7.352.0=pypi_0
olefile=0.46=py37_0
openssl=1.1.1d=he774522_3
pillow=6.2.1=py37hdc69c19_0
pip=19.3.1=py37_0
pip-conflict-checker=0.6.0=pypi_0
plac=1.1.3=pypi_0
preshed=3.0.2=pypi_0
pycparser=2.19=py37_0
pyflakes=2.1.1=pypi_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.5=py_0
pyqt=5.9.2=py37h6538335_2
pysocks=1.7.1=py37_0
python=3.7.5=h8c8aaf0_0
python-dateutil=2.8.0=pypi_0
pytorch=1.3.1=py3.7_cuda101_cudnn7_0
pytz=2019.3=py_0
pyyaml=5.1.2=pypi_0
qt=5.9.7=vc14h73c81de_0
regex=2019.11.1=pypi_0
requests=2.22.0=py37_1
s3transfer=0.2.1=pypi_0
sacremoses=0.0.35=pypi_0
scikit-learn=0.21.3=py37h6288b17_0
scipy=1.3.2=pypi_0
sentencepiece=0.1.83=pypi_0
setuptools=42.0.2=py37_0
sip=4.19.8=py37h6538335_0
six=1.12.0=pypi_0
soupsieve=1.9.5=py37_0
spacy=2.2.3=pypi_0
sqlite=3.30.1=he774522_0
srsly=0.2.0=pypi_0
termcolor=1.1.0=pypi_0
thinc=7.3.1=pypi_0
tk=8.6.8=hfa6e2cd_0
torchvision=0.4.2=py37_cu101
tornado=6.0.3=py37he774522_0
tqdm=4.38.0=pypi_0
transformers=2.3.0=pypi_0
urllib3=1.25.7=py37_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_0
wasabi=0.4.0=pypi_0
wheel=0.33.6=py37_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
xz=5.2.4=h2fa13f4_4
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
